refactor: relocate API URL construction to config
- Moved the `getApiUrl` function from the lib directory to a new `api.ts` file in the config directory, enhancing code organization and maintainability.
- Updated import paths in `steam-guide-card.tsx` to reflect this change.
