+++ src/components/steam-guide/steam-guide-card.tsx
@@ -4,7 +4,7 @@ import { ArrowRightIcon, BookHeartIcon, EyeIcon, StarIcon } from "lucide-react";
 import millify from "millify";
 
 import type { SteamGuideResponse } from "@/types/steam";
-import { getApiUrl } from "@/lib/api";
+import { getApiUrl } from "@/config/api";
 import { steamConfig } from "@/lib/constants";
 import { Button } from "@/components/ui/button";
 import {
